# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * InfoFi Labz API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Kol {
  id: number;
  name: string;
  followers: string;
  niche: string;
  imageUrl: string;
  twitter?: string | null;
  telegram?: string | null;
}

export interface KolInput {
  name: string;
  followers: string;
  niche: string;
  imageUrl: string;
  twitter?: string;
  telegram?: string;
}

export interface FormField {
  id: string;
  type: "text" | "textarea" | "url" | "number" | "select";
  label: string;
  placeholder?: string;
  required: boolean;
  options?: string[];
}

export interface Campaign {
  id: number;
  title: string;
  tag: string;
  description: string;
  applyLink: string;
  imageUrl: string;
  type: string;
  /** active, closed, or pending */
  status: string;
  tagIds: number[];
  formFields: FormField[];
}

export interface CampaignInput {
  title: string;
  tag: string;
  description: string;
  applyLink: string;
  imageUrl: string;
  type: string;
  /** active, closed, or pending */
  status: string;
  tagIds: number[];
  formFields?: FormField[];
}

export interface CampaignSubmission {
  id: number;
  campaignId: number;
  userId: number;
  answers: Record<string, string>;
  /** pending, approved, or rejected */
  status: string;
  createdAt: string;
  campaignTitle?: string | null;
  campaignType?: string | null;
  campaignImageUrl?: string | null;
  userLogin?: string | null;
}

export interface Case {
  id: number;
  project: string;
  result: string;
  category: string;
  imageUrl: string;
}

export interface CaseInput {
  project: string;
  result: string;
  category: string;
  imageUrl: string;
}

export interface AuthInput {
  login: string;
  password: string;
}

export interface AuthUser {
  id: number;
  login: string;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Application {
  id: number;
  /** kol or project */
  type: string;
  name: string;
  telegram: string;
  socialMedia: string;
  additionalLink1?: string | null;
  additionalLink2?: string | null;
  about: string;
  /** pending, verified, or rejected */
  status: string;
  userId?: number | null;
  createdAt: string;
}

export interface ApplicationInput {
  /** kol or project */
  type: string;
  name: string;
  telegram: string;
  socialMedia: string;
  additionalLink1?: string;
  additionalLink2?: string;
  about: string;
  userId?: number | null;
}

export interface Section {
  id: number;
  name: string;
  slug: string;
  icon: string;
  description: string;
  color: string;
  sortOrder: number;
  isActive: boolean;
}

export interface SectionInput {
  name: string;
  slug: string;
  icon: string;
  description: string;
  color: string;
  sortOrder: number;
  isActive: boolean;
}

export interface Tag {
  id: number;
  name: string;
  slug: string;
  sectionId?: number | null;
  color: string;
}

export interface TagInput {
  name: string;
  slug: string;
  sectionId?: number | null;
  color: string;
}

export interface ApplicationStatusUpdate {
  /** verified or rejected */
  status: string;
}

export interface UserProfile {
  id: number;
  login: string;
  rank: string;
  evmWallet?: string | null;
  bio?: string | null;
  createdAt: string;
}

export interface UserProfileUpdate {
  evmWallet?: string | null;
  bio?: string | null;
}

export interface UserRankUpdate {
  /** beginner, advanced, elite, or early */
  rank: string;
}

export interface Post {
  id: number;
  title: string;
  content: string;
  imageUrl?: string | null;
  createdAt: string;
}

export interface PostInput {
  title: string;
  content: string;
  imageUrl?: string | null;
}

export interface DeleteResult {
  success: boolean;
}

export type ListCampaignsParams = {
  type?: string;
};

export type ListTagsParams = {
  sectionId?: number;
};